Novas classes para exceções personalidas!

+++ src/main/java/com/vsm/devcase/cliente/ClienteService.java
@@ -4,12 +4,15 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+
+import com.vsm.devcase.exception.DevcaseException;
 
 @Service
 public class ClienteService {
 
-	@Autowired	
+	@Autowired
 	private ClienteRepository clienteRepository;
 
 	public Page<Cliente> buscarTodosOsClientes(Pageable pageable) {
@@ -23,10 +26,13 @@ public class ClienteService {
 
 	public Cliente buscarCliente(Integer codigo) {
 		Cliente cliente = clienteRepository.findOne(codigo);
+		if (cliente == null) {
+			throw new DevcaseException("Cliente não encontrado!", HttpStatus.NOT_FOUND);
+		}
 		return cliente;
 	}
 
-	public Cliente atualizarCliente(Integer codigo,  Cliente cliente) {
+	public Cliente atualizarCliente(Integer codigo, Cliente cliente) {
 		Cliente clienteSalvo = clienteRepository.findOne(codigo);
 		BeanUtils.copyProperties(cliente, clienteSalvo, "codigo");
 		return clienteRepository.save(clienteSalvo);
@@ -34,6 +40,9 @@ public class ClienteService {
 
 	public void removerCliente(Integer codigo) {
 		Cliente cliente = clienteRepository.findOne(codigo);
+		if (cliente == null) {
+			throw new DevcaseException("Cliente não encontrado!", HttpStatus.NOT_FOUND);
+		}
 		clienteRepository.delete(cliente);
 	}
